Fixes the following issues:
#114 Logout before account creation doesn't log you out correctly
#113 Saving navigation structure converts "link" field to lowercase

Also adds a _DATABASE_CHANGES.txt file to be temporarily used as a place to keep notes on any database changes that need to happen for the next incremental version of the framework.  This will go away as soon as we have a database deployment system in place.

Also refactored the Backup module.